multimodal-rag-demo-main/requirements.api.txt
# Core dependencies
fastapi==0.105.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyYAML==6.0.1

# Database and caching
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Background tasks
celery==5.3.4
kombu==5.3.4

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI and ML dependencies
cohere==5.0.0
google-generativeai==0.7.0
numpy==1.26.3
scikit-learn==1.4.0
faiss-cpu==1.7.4

# Document processing
pdf2image==1.17.0
PyPDF2==3.0.1
pillow==10.2.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# HTTP client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
